MOD: network for version 2.0
